+++ include/GafferUI/GraphGadget.h
@@ -201,6 +201,8 @@ class GAFFERUI_API GraphGadget : public ContainerGadget
 		void rootChildRemoved( Gaffer::GraphComponent *root, Gaffer::GraphComponent *child );
 		void selectionMemberAdded( Gaffer::Set *set, IECore::RunTimeTyped *member );
 		void selectionMemberRemoved( Gaffer::Set *set, IECore::RunTimeTyped *member );
+		void focusMemberAdded( Gaffer::Set *set, IECore::RunTimeTyped *member );
+		void focusMemberRemoved( Gaffer::Set *set, IECore::RunTimeTyped *member );
 		void filterMemberAdded( Gaffer::Set *set, IECore::RunTimeTyped *member );
 		void filterMemberRemoved( Gaffer::Set *set, IECore::RunTimeTyped *member );
 		void inputChanged( Gaffer::Plug *dstPlug );
@@ -255,6 +257,8 @@ class GAFFERUI_API GraphGadget : public ContainerGadget
 		boost::signals::scoped_connection m_rootChildRemovedConnection;
 		boost::signals::scoped_connection m_selectionMemberAddedConnection;
 		boost::signals::scoped_connection m_selectionMemberRemovedConnection;
+		boost::signals::scoped_connection m_focusMemberAddedConnection;
+		boost::signals::scoped_connection m_focusMemberRemovedConnection;
 
 		Gaffer::SetPtr m_filter;
 		boost::signals::scoped_connection m_filterMemberAddedConnection;